Create a Health System.
I moved the Health System from the GameManager to its own class so I can use the code in other scripts.
