+++ Assets/Jacob/Scripts/Controllers/GameManager.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using Jacob.Scripts.Data;
+﻿using Jacob.Scripts.Data;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -19,11 +18,7 @@ namespace Jacob.Scripts.Controllers
 		[SerializeField] [HideInInspector] public UnityEvent<float> whenMoneyChangesFloat;
 		[SerializeField] [HideInInspector] public UnityEvent<string> whenMoneyChangesString;
 
-		/// <summary>
-		/// A Health stat.
-		/// </summary>
-		internal double Health { get; private set; }
-
+		internal HealthSystem Health;
 		internal CashSystem Cash;
 
 		/// <summary>
@@ -39,7 +34,10 @@ namespace Jacob.Scripts.Controllers
 
 		private void InitializeStats()
 		{
-			Health = maxHealth;
+			Health = new HealthSystem(maxHealth)
+			{
+				whenYouDie = () => whenYouDie.Invoke()
+			};
 			Cash = new CashSystem
 			{
 				SetAction = SetAction
@@ -59,35 +57,7 @@ namespace Jacob.Scripts.Controllers
 			}
 		}
 
-		/// <summary>
-		/// Add some Health to the Health stat. Will clamp to the maxHealth value as you can't go over your maxHealth.
-		/// </summary>
-		/// <param name="amountOfHealth">The amount of Health you want to add.</param>
-		public void AddHealth(double amountOfHealth)
-		{
-			Health += Math.Clamp(amountOfHealth, 0, maxHealth);
-			if (Health <= 0) Kill();
-		}
-
-		/// <summary>
-		/// Subtract some Health to the Health stat. Will clamp to the maxHealth value as you can't go under your maxHealth.
-		/// </summary>
-		/// <param name="amountOfHealth"></param>
-		public void SubtractHealth(double amountOfHealth)
-		{
-			Health -= Math.Clamp(amountOfHealth, 0, maxHealth);
-			if (Health <= 0) Kill();
-		}
-
-		/// <summary>
-		/// When you completely run out of Health.
-		/// Runs a UnityEvent so you can run whatever you want when you run out of Health.
-		/// </summary>
-		public void Kill()
-		{
-			if (Health != 0) Health = 0;
-			whenYouDie.Invoke();
-		}
+		
 
 		/// <summary>
 		/// Get the active GameManager. Adapts to the name of the GameObject your GameManager is on.
